feat(threatscore): restore API threatscore snapshots
